backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
python-dotenv
tzdata
google-api-python-client>=2.126.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
